# apps/backend/src/shared/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mandarin Learning Platform API
  version: 1.0.0
  description: |
    Production backend API for Mandarin vocabulary learning and AI-powered
    conversation generation. All endpoints are versioned under /api/v1.

    This reference documents the full NestJS 11 route surface: 54 route paths
    across the 15 backend modules (audio, auth, characters, chengyu,
    foundations, grammar, health, mnemonics, phonetic-clusters, progression,
    quiz, radicals, readers, review, words — pinyin lives under foundations).

    Authentication model (per route):
      - Public — no auth required (static reference data, health, auth
        register/login/refresh/logout).
      - Required auth — `Authorization: Bearer <access_token>`; user-scoped
        state or AI/vendor-cost generation (`RequireAuthGuard`).
      - Optional auth — bearer token is best-effort (`OptionalAuthGuard`): an
        absent or invalid token proceeds as a guest (`req.userId` undefined),
        never 401. Guest responses are session-local/empty shapes.

    All 4xx/5xx errors use the `{ code, message, requestId }` envelope.
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://mandarin-vite-react-ts-production.up.railway.app/api
    description: Production server (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HTTP-only refresh token cookie
  schemas:
    ErrorEnvelope:
      type: object
      description: |
        Standard error envelope — the global 24-3 `AppExceptionFilter`
        serializes every thrown `HttpException` into this shape.
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request ID (echoes the `X-Request-Id` response header)
      required:
        - code
        - message
        - requestId
...
paths:
...
  /v1/pinyin/search:
    get:
      summary: Search characters by pinyin
      tags:
        - Pinyin
      description: |
        Searches characters by pinyin query. Public. Unlike the characters
        search, this route is NOT shadowed (the pinyin router registered only
        `/search`), so it returns a real 2xx.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Pinyin search query (required)
        - name: tone
          in: query
          required: false
          schema:
            type: integer
          description: Tone filter (1-5)
        - name: page
          in: query
          required: false
          schema:
            type: integer
          description: Page number (default 1)
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
          description: Page size (default 50)
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Missing q or invalid tone (VALIDATION_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to search pinyin (INTERNAL_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ══════════════════════════════════════════════════════════════════
  # Grammar
  # ══════════════════════════════════════════════════════════════════
...